support filter by if a key is expired
