+++ src/RDBCli/Helpers/CommonHelper.Const.cs
@@ -2,7 +2,7 @@
 
 namespace RDBCli
 {
-    internal static class CommonHelper
+    internal static partial class CommonHelper
     {
         // compress Tpl/tpl.html to get this value.
         internal const string TplHtmlString = "<!DOCTYPE html><html lang=\"en\"><head><meta charset=\"UTF-8\"><meta http-equiv=\"X-UA-Compatible\"content=\"IE=edge\"><meta name=\"viewport\"content=\"width=device-width, initial-scale=1.0\"><title>rdb-cli,offline key analysis</title><link rel=\"stylesheet\"href=\"https://unpkg.com/bootstrap@5.1.3/dist/css/bootstrap.min.css\"/><link rel=\"stylesheet\"href=\"https://unpkg.com/bootstrap-table@1.19.1/dist/bootstrap-table.min.css\"><style>.bootstrap-table.fixed-table-pagination>.pagination-detail.page-list{display:none}</style><script src=\"https://unpkg.com/jquery@3.6.0/dist/jquery.min.js\"></script><script src=\"https://unpkg.com/bootstrap@5.1.3/dist/js/bootstrap.min.js\"></script><script src=\"https://unpkg.com/chart.js@3.7.1/dist/chart.min.js\"></script><script src=\"https://unpkg.com/bootstrap-table@1.19.1/dist/bootstrap-table.min.js\"></script></head><body><nav class=\"navbar navbar-expand-sm bg-info navbar-dark\"><div class=\"ms-auto me-auto\"><a class=\"navbar-brand fs-4\"href=\"https://github.com/catcherwong/rdb-tools\">rdb-cli</a></div></nav><div class=\"container\"><br/><div class=\"page-header border-start border-light border-3\"><h3>Basic Information</h3></div><br/><div class=\"row\"><div class=\"col-md-4\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div class=\"lead display-7\">RDB Version:<b id=\"rdbVer\"></b></div><div class=\"lead display-7\">Redis Version:<b id=\"redisVer\"></b></div></div></div></div><div class=\"col-md-4\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"totalKeyBytes\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Total Memory Usage of Keys</p></div></div></div><div class=\"col-md-4\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"totalKeyNum\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Total Number of Keys</p></div></div></div></div><br/><div class=\"row\"><div class=\"col-md-3\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"expKeyNum\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Expired Keys</p></div></div></div><div class=\"col-md-3\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"expKeyBytes\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Memory Usage by Expired Keys</p></div></div></div><div class=\"col-md-3\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"permKeyNum\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Permanent Keys</p></div></div></div><div class=\"col-md-3\"><div class=\"card\"style=\"height: 100px;\"><div class=\"card-body text-center\"><div id=\"permKeyBytes\"class=\"lead text-primary display-6\"></div><p class=\"small text-muted\">Memory Usage by Permanent Keys</p></div></div></div></div><br/><div class=\"page-header border-start border-light border-3\"><h3>Details Information</h3></div><br/><div class=\"row\"><div class=\"col-md-6\"><canvas id=\"typeNum\"width=\"400\"height=\"400\"></canvas></div><div class=\"col-md-6\"><canvas id=\"typeByte\"width=\"400\"height=\"400\"></canvas></div></div><br/><br/><div class=\"row\"><div class=\"col-md-6\"><canvas id=\"expiryNum\"width=\"400\"height=\"400\"></canvas></div><div class=\"col-md-6\"><canvas id=\"expiryByte\"width=\"400\"height=\"400\"></canvas></div></div><br/><br/><div class=\"row\"><div class=\"col-md-6\"><canvas id=\"idleOrFreqNum\"width=\"400\"height=\"400\"></canvas></div><div class=\"col-md-6\"><canvas id=\"idleOrFreqByte\"width=\"400\"height=\"400\"></canvas></div></div><br/><br/><div class=\"row\"><div class=\"col-md-12\"><ul class=\"nav nav-tabs\"role=\"tablist\"><li class=\"nav-item\"><a class=\"nav-link active\"data-bs-toggle=\"tab\"href=\"#topPrefix\">Top Key Prefixes</a></li><li class=\"nav-item\"><a class=\"nav-link\"data-bs-toggle=\"tab\"href=\"#topBigKeys\">Top Big Keys</a></li><li class=\"nav-item\"id=\"tbStreams\"><a class=\"nav-link\"data-bs-toggle=\"tab\"href=\"#streams\">Top Streams</a></li><li class=\"nav-item\"id=\"tbFunctions\"><a class=\"nav-link\"data-bs-toggle=\"tab\"href=\"#functions\">Functions</a></li></ul><div class=\"tab-content\"><div id=\"topBigKeys\"class=\"container tab-pane fade\"><br><table id=\"topBigKeysTable\"></table></div><div id=\"topPrefix\"class=\"container tab-pane active\"><br><table id=\"topPrefixTable\"></table></div><div id=\"functions\"class=\"container tab-pane fade\"><br><table id=\"functionTable\"></table></div><div id=\"streams\"class=\"container tab-pane fade\"><br><table id=\"streamsTable\"></table></div></div></div></div></div>" +
@@ -11,193 +11,5 @@ namespace RDBCli
 
         internal const string AlreadyExpired = "Already Expired";
         internal const string Permanent = "Permanent";
-
-        /// <summary>
-        /// Get a fuzzy redis version from a rdb version.
-        /// Mainly for not get `redis-ver` from aux field.
-        /// </summary>
-        /// <param name="rdbVer">RDB VERSION </param>
-        /// <returns>REDIS VERSION</returns>
-        internal static string GetFuzzyRedisVersion(int rdbVer)
-        {
-            var ver = "unknow";
-
-            if (rdbVer == 11)
-            {
-                ver = "7.2.x";
-            }
-            else if (rdbVer == 10)
-            {
-                ver = "7.0.x";
-            }
-            else if (rdbVer == 9)
-            {
-                // 5.0.0 ~ 5.0.14 ~ 6.2.6
-                ver = "5.x";
-            }
-            else if (rdbVer == 8)
-            {
-                // 4.0.0 ~ 4.0.14
-                ver = "4.0.x";
-            }
-            else if (rdbVer == 7)
-            {
-                // 3.2.0 ~ 3.2.13
-                ver = "3.2.x";
-            }
-            else if (rdbVer == 6)
-            {
-                // 2.6.0 ~ 2.8.24 ~ 3.0.7
-                ver = "2.8.x";
-            }
-            else if (rdbVer <= 5)
-            {
-                ver = "2.x";
-            }
-
-            return ver;
-        }
-
-        internal static string GetExpireString(long exp)
-        {
-            var res = exp.ToString();
-
-            if (exp > 0)
-            {
-                var sub = DateTimeOffset.FromUnixTimeMilliseconds(exp).Subtract(DateTimeOffset.UtcNow);
-
-                // 0~1h, 1~3h, 3~12h, 12~24h, 24~72h, 72~168h, 168h~
-                var hour = sub.TotalHours;
-                if (hour <= 0)
-                {
-                    res = AlreadyExpired;
-                }
-                else if (hour > 0 && hour < 1)
-                {
-                    res = "0~1h";
-                }
-                else if (hour >= 1 && hour < 3)
-                {
-                    res = "1~3h";
-                }
-                else if (hour >= 3 && hour < 12)
-                {
-                    res = "3~12h";
-                }
-                else if (hour >= 12 && hour < 24)
-                {
-                    res = "12~24h";
-                }
-                else if (hour >= 24 && hour < 72)
-                {
-                    res = "1~3d";
-                }
-                else if (hour >= 72 && hour < 168)
-                {
-                    res = "3~7d";
-                }
-                else if (hour >= 168)
-                {
-                    res = ">7d";
-                }
-            }
-            else if (exp == 0)
-            {
-                res = Permanent;
-            }
-
-            return res;
-        }
-
-        internal static string GetIdleString(ulong idle)
-        {
-            var res = idle.ToString();
-
-            // 0~1h = 3600, 1~3h = 10800, 3~12h = 43200, 12~24h = 86400, 1~3d = 259200, 3~7d = 604800
-
-            if (idle < 3600)
-            {
-                res = "0~1h";
-            }
-            else if (idle >= 3600 && idle < 10800)
-            {
-                res = "1~3h";
-            }
-            else if (idle >= 10800 && idle < 43200)
-            {
-                res = "3~12h";
-            }
-            else if (idle >= 43200 && idle < 86400)
-            {
-                res = "12~24h";
-            }
-            else if (idle >= 86400 && idle < 259200)
-            {
-                res = "1~3d";
-            }
-            else if (idle >= 259200 && idle < 604800)
-            {
-                res = "3~7d";
-            }
-            else if (idle >= 604800)
-            {
-                res = ">7d";
-            }
-
-            return res;
-        }
-
-        internal static string GetFreqString(int freq)
-        {
-            // 0~50, 50~100, 100~150, 150~200, 200~250, 250~
-            var res = freq.ToString();
-
-            if (freq < 50)
-            {
-                res = "0~50";
-            }
-            else if (freq >= 50 && freq < 100)
-            {
-                res = "50~100";
-            }
-            else if (freq >= 100 && freq < 150)
-            {
-                res = "100~150";
-            }
-            else if (freq >= 150 && freq < 200)
-            {
-                res = "150~200";
-            }
-            else if (freq >= 200 && freq < 250)
-            {
-                res = "200~250";
-            }
-            else if (freq >= 250)
-            {
-                res = ">250";
-            }
-
-            return res;
-
-        }
-
-        internal const char SplitChar = ':';
-        internal static string GetShortKey(string key)
-        {
-            var len = key.Length;
-
-            if (len > 1024)
-            {
-                var span = key.AsSpan();
-
-                var b = span.Slice(0, 10).ToString();
-                var e = span.Slice(len - 6, 5).ToString();
-
-                var n = $"{b}...({len - 15} more bytes)...{e}";
-                return n;
-            }
-
-            return key;
-        }
     }
 }
